+++ pom.xml
@@ -11,7 +11,7 @@
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
 
-  <groupId>co.eightwelve</groupId>
+  <groupId>co.eightwelve.cdt</groupId>
   <artifactId>cdt-be</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <name>cdt-be</name>
