+++ src/main/java/co/eightwelve/cdt/be/CdtBeApplication.java
@@ -1,4 +1,4 @@
-package co.eightwelve.be;
+package co.eightwelve.cdt.be;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
