+++ src/main/java/co/eightwelve/cdt/be/infrastructure/security/CorsConfig.java
@@ -1,4 +1,4 @@
-package co.eightwelve.be.infrastructure.security;
+package co.eightwelve.cdt.be.infrastructure.security;
 
 import java.util.Collections;
 import java.util.List;
