+++ src/test/java/co/eightwelve/cdt/be/infrastructure/swagger/SwaggerIntTest.java
@@ -1,4 +1,4 @@
-package co.eightwelve.be.infrastructure.actuator;
+package co.eightwelve.cdt.be.infrastructure.swagger;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -8,7 +8,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 /**
@@ -17,14 +19,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class ActuatorIntTest {
+public class SwaggerIntTest {
 
   @Autowired
   private MockMvc mockMvc;
 
   @Test
-  public void providesHealthEndpoint() throws Exception {
-    mockMvc.perform(get("/actuator/health"))
-        .andExpect(status().isOk());
+  public void providesSwaggerApiDocumentation() throws Exception {
+    mockMvc.perform(get("/v2/api-docs"))
+        .andExpect(status().isOk())
+        .andExpect(content().string(containsString("Countdown-together")));
   }
 }
